Backend/requirements.txt
flask
flask-cors
langdetect
sentence-transformers
scikit-learn
requests
python-dotenv
firebase-admin
pillow
pytesseract
